Refactor tank logic into TankService and module
Moved all tank-related logic from GameService to a new TankService in the tank module. GameService now delegates tank operations such as piece placement, water quality calculation, adjacency checks, and consumable processing to TankService, improving separation of concerns and maintainability.
