+++ game-engine/src/game/game.module.ts
@@ -5,9 +5,10 @@ import { PlayerModule } from '../player/player.module';
 import { AIModule } from '../ai/ai.module';
 import { BattleModule } from '../battle/battle.module';
 import { EconomyModule } from '../economy/economy.module';
+import { TankModule } from '../tank/tank.module';
 
 @Module({
-  imports: [PlayerModule, AIModule, BattleModule, EconomyModule], // Import PlayerModule, AIModule, BattleModule, and EconomyModule
+  imports: [PlayerModule, AIModule, BattleModule, EconomyModule, TankModule], // Import all service modules
   providers: [GameService, GameGateway],
   exports: [GameService],
 })
