+++ game-engine/src/game/game.service.ts
@@ -6,6 +6,7 @@ import { PlayerService } from '../player/player.service';
 import { AIService } from '../ai/ai.service';
 import { BattleService } from '../battle/battle.service';
 import { EconomyService } from '../economy/economy.service';
+import { TankService } from '../tank/tank.service';
 
 // Game constants
 const MAX_ROUNDS = 15;
@@ -21,7 +22,8 @@ export class GameService {
     private playerService: PlayerService,
     private aiService: AIService,
     private battleService: BattleService,
-    private economyService: EconomyService
+    private economyService: EconomyService,
+    private tankService: TankService
   ) {}
 
   async createSession(socketId: string, playerId: string): Promise<GameState> {
@@ -139,7 +141,7 @@ export class GameService {
     });
 
     // Recalculate water quality after adding new piece
-    gameState.playerTank.waterQuality = this.calculateWaterQuality(gameState.playerTank);
+    gameState.playerTank.waterQuality = this.tankService.calculateWaterQuality(gameState.playerTank);
 
     this.updateGameState(socketId, gameState);
     return gameState;
@@ -165,7 +167,7 @@ export class GameService {
     
     // Remove piece from tank and grid
     gameState.playerTank.pieces.splice(pieceIndex, 1);
-    this.removePieceFromGrid(gameState.playerTank, piece);
+    this.tankService.removePieceFromGrid(gameState.playerTank, piece);
 
     // Add gold transaction
     gameState.goldHistory.push({
@@ -180,7 +182,7 @@ export class GameService {
     });
 
     // Recalculate water quality after removing piece
-    gameState.playerTank.waterQuality = this.calculateWaterQuality(gameState.playerTank);
+    gameState.playerTank.waterQuality = this.tankService.calculateWaterQuality(gameState.playerTank);
 
     this.updateGameState(socketId, gameState);
     return gameState;
@@ -236,27 +238,11 @@ export class GameService {
       throw new BadRequestException('Piece not found');
     }
 
-    if (action === 'remove') {
-      this.removePieceFromGrid(gameState.playerTank, piece);
-      piece.position = undefined;
-    } else {
-      // Validate position
-      if (!this.isValidPosition(gameState.playerTank, piece, position)) {
-        throw new BadRequestException('Invalid position for piece');
-      }
-
-      // Remove from old position
-      if (piece.position) {
-        this.removePieceFromGrid(gameState.playerTank, piece);
-      }
-
-      // Place at new position
-      piece.position = position;
-      this.placePieceOnGrid(gameState.playerTank, piece);
+    try {
+      this.tankService.updateTankPiece(gameState.playerTank, pieceId, position, action);
+    } catch (error) {
+      throw new BadRequestException(error.message);
     }
-
-    // Recalculate water quality after placing/moving/removing pieces
-    gameState.playerTank.waterQuality = this.calculateWaterQuality(gameState.playerTank);
 
     this.updateGameState(socketId, gameState);
     return gameState;
@@ -482,8 +468,8 @@ export class GameService {
     gameState.battleState = undefined;
     
     // Respawn all dead fish and plants (but not consumables)
-    this.respawnPieces(gameState.playerTank);
-    this.respawnPieces(gameState.opponentTank);
+    this.tankService.respawnPieces(gameState.playerTank);
+    this.tankService.respawnPieces(gameState.opponentTank);
     
     // Shop should already be generated from finalizeBattleRewards, but ensure it exists
     if (!gameState.shop || gameState.shop.length === 0) {
@@ -645,7 +631,7 @@ export class GameService {
     const placedPieces = gameState.playerTank.pieces.filter(p => p.position);
     
     placedPieces.forEach(piece => {
-      result[piece.id] = this.calculatePieceStats(piece, placedPieces);
+      result[piece.id] = this.tankService.calculatePieceStats(piece, placedPieces);
     });
     
     return result;
@@ -662,11 +648,11 @@ export class GameService {
     // Only add new pieces to existing tank, don't regenerate from scratch
     const { remainingGold } = this.aiService.updateOpponentTank(
       gameState,
-      (tank) => this.calculateWaterQuality(tank),
-      (tank, piece) => this.findValidPositionForOpponent(tank, piece),
-      (tank, piece) => this.placePieceOnGrid(tank, piece),
-      (tank, piece) => this.removePieceFromGrid(tank, piece),
-      (tank, piece, targetType) => this.findOptimalConsumablePosition(tank, piece)
+      (tank) => this.tankService.calculateWaterQuality(tank),
+      (tank, piece) => this.tankService.findValidPositionForOpponent(tank, piece),
+      (tank, piece) => this.tankService.placePieceOnGrid(tank, piece),
+      (tank, piece) => this.tankService.removePieceFromGrid(tank, piece),
+      (tank, piece, targetType) => this.tankService.findOptimalConsumablePosition(tank, piece)
     );
     gameState.opponentGold = remainingGold; // Track remaining gold
     gameState.phase = 'placement';
@@ -683,11 +669,11 @@ export class GameService {
     }
 
     // Process consumables before battle starts
-    this.processConsumables(gameState.playerTank);
-    this.processConsumables(gameState.opponentTank);
+    this.tankService.processConsumables(gameState.playerTank);
+    this.tankService.processConsumables(gameState.opponentTank);
 
     // Initialize battle state but keep on placement phase
-    const battleState = this.battleService.initializeBattleState(gameState, (piece, pieces) => this.calculatePieceStats(piece, pieces));
+    const battleState = this.battleService.initializeBattleState(gameState, (piece, pieces) => this.tankService.calculatePieceStats(piece, pieces));
     gameState.battleState = battleState;
     gameState.phase = 'battle'; // Still need to track battle state internally
     
@@ -721,439 +707,6 @@ export class GameService {
     const playerId = this.playerService.getPlayerIdFromSocket(socketId);
     this.playerService.updateSession(playerId, gameState);
   }
-
-
-
-  private isValidPosition(tank: Tank, piece: GamePiece, position: Position): boolean {
-    for (const offset of piece.shape) {
-      const x = position.x + offset.x;
-      const y = position.y + offset.y;
-      
-      if (x < 0 || x >= 8 || y < 0 || y >= 6) {
-        return false;
-      }
-      
-      if (tank.grid[y][x] && tank.grid[y][x] !== piece.id) {
-        return false;
-      }
-    }
-    
-    return true;
-  }
-
-  private removePieceFromGrid(tank: Tank, piece: GamePiece): void {
-    if (!piece.position) return;
-    
-    for (const offset of piece.shape) {
-      const x = piece.position.x + offset.x;
-      const y = piece.position.y + offset.y;
-      
-      if (x >= 0 && x < 8 && y >= 0 && y < 6) {
-        tank.grid[y][x] = null;
-      }
-    }
-  }
-
-  private placePieceOnGrid(tank: Tank, piece: GamePiece): void {
-    if (!piece.position) return;
-    
-    for (const offset of piece.shape) {
-      const x = piece.position.x + offset.x;
-      const y = piece.position.y + offset.y;
-      
-      if (x >= 0 && x < 8 && y >= 0 && y < 6) {
-        tank.grid[y][x] = piece.id;
-      }
-    }
-  }
-
-  private processConsumables(tank: Tank): void {
-    // Find all consumables that are placed on the grid
-    const consumables = tank.pieces.filter(p => p.type === 'consumable' && p.position);
-    
-    if (consumables.length === 0) return;
-    
-    console.log(`🍽️ Processing ${consumables.length} consumables for tank ${tank.id}`);
-    
-    // For each consumable, apply bonuses to adjacent fish
-    for (const consumable of consumables) {
-      if (!consumable.position) continue;
-      
-      // Get all adjacent positions for this multi-cell consumable
-      const adjacentPositions = this.getAdjacentPositionsForPiece(consumable);
-      
-      // Find adjacent fish - check if any of their cells are adjacent to the consumable
-      const adjacentFish = tank.pieces.filter(p => {
-        if (p.type !== 'fish' || !p.position) return false;
-        
-        // Check if any cell of the fish is adjacent to the consumable
-        return p.shape.some(offset => {
-          const fishCellPos = { x: p.position!.x + offset.x, y: p.position!.y + offset.y };
-          return adjacentPositions.some(adjPos => adjPos.x === fishCellPos.x && adjPos.y === fishCellPos.y);
-        });
-      });
-      
-      // Apply permanent bonuses to each adjacent fish
-      for (const fish of adjacentFish) {
-        // Initialize permanent bonuses if not exists
-        if (!fish.permanentBonuses) {
-          fish.permanentBonuses = {
-            attack: 0,
-            health: 0,
-            speed: 0,
-            sources: []
-          };
-        }
-        
-        // Apply bonuses
-        const attackBonus = consumable.attackBonus || 0;
-        const healthBonus = consumable.healthBonus || 0;
-        const speedBonus = consumable.speedBonus || 0;
-        
-        fish.permanentBonuses.attack += attackBonus;
-        fish.permanentBonuses.health += healthBonus;
-        fish.permanentBonuses.speed += speedBonus;
-        
-        // Track the source
-        const existingSource = fish.permanentBonuses.sources.find(s => s.name === consumable.name);
-        if (existingSource) {
-          existingSource.count++;
-        } else {
-          fish.permanentBonuses.sources.push({
-            name: consumable.name,
-            count: 1,
-            attackBonus,
-            healthBonus,
-            speedBonus
-          });
-        }
-        
-        // Don't modify base stats - only track in permanentBonuses
-        console.log(`🎯 ${consumable.name} applied permanent bonuses to ${fish.name}: +${attackBonus} ATK, +${healthBonus} HP, +${speedBonus} SPD`);
-      }
-      
-      // Remove consumable from grid
-      this.removePieceFromGrid(tank, consumable);
-    }
-    
-    // Remove all consumables from the tank pieces array
-    tank.pieces = tank.pieces.filter(p => p.type !== 'consumable' || !p.position);
-  }
-
-  private calculateWaterQuality(tank: Tank): number {
-    // Start with the tank's original baseline quality (never changes)
-    let quality = tank.baseWaterQuality;
-    
-    // Count placed pieces and calculate the total effect
-    let totalEffect = 0;
-    let fishCount = 0;
-    let plantCount = 0;
-    let filterCount = 0;
-    
-    for (const piece of tank.pieces) {
-      if (!piece.position) continue; // Only count placed pieces
-      
-      
-      if (piece.type === 'fish') {
-        totalEffect -= 1; // Fish decrease quality
-        fishCount++;
-      } else if (piece.type === 'plant') {
-        totalEffect += 1; // Plants increase quality
-        plantCount++;
-      } else if (piece.type === 'equipment') {
-        // Most equipment is neutral, but check for special items
-        if (piece.name === 'Sponge Filter' || piece.tags?.includes('filter')) {
-          totalEffect += 1; // Filters improve quality
-          filterCount++;
-        }
-      }
-    }
-    
-    
-    // Apply the effect to the starting quality
-    quality += totalEffect;
-    
-    
-    // Clamp quality between 1-10
-    return Math.max(1, Math.min(10, quality));
-  }
-
-  private respawnPieces(tank: Tank): void {
-    // Respawn all non-consumable pieces (fish, plants, equipment), preserving their permanent bonuses
-    for (const piece of tank.pieces) {
-      if (piece.type === 'fish' || piece.type === 'plant' || piece.type === 'equipment') {
-        // Clear the dead flag
-        (piece as any).isDead = false;
-        
-        // Restore health to base max (permanent bonuses are applied separately in calculations)
-        const baseMaxHealth = PIECE_LIBRARY.find(p => p.name === piece.name)?.stats.maxHealth || piece.stats.maxHealth;
-        piece.stats.maxHealth = baseMaxHealth;
-        piece.stats.health = baseMaxHealth;
-        
-        const permanentHealthBonus = piece.permanentBonuses?.health || 0;
-        console.log(`🔄 Respawned ${piece.name} with ${piece.stats.health}/${piece.stats.maxHealth} HP base (+${permanentHealthBonus} permanent bonus tracked separately)`);
-      }
-    }
-  }
-
-  private aretwoPiecesAdjacent(piece1: GamePiece, piece2: GamePiece): boolean {
-    if (!piece1.position || !piece2.position || piece1.id === piece2.id) return false;
-    
-    // Get all cells occupied by piece1
-    const piece1Cells = piece1.shape.map(offset => ({
-      x: piece1.position!.x + offset.x,
-      y: piece1.position!.y + offset.y
-    }));
-    
-    // Get all cells occupied by piece2
-    const piece2Cells = piece2.shape.map(offset => ({
-      x: piece2.position!.x + offset.x,
-      y: piece2.position!.y + offset.y
-    }));
-    
-    // Check if any cell of piece1 is adjacent to any cell of piece2
-    return piece1Cells.some(cell1 => {
-      return piece2Cells.some(cell2 => {
-        const dx = Math.abs(cell1.x - cell2.x);
-        const dy = Math.abs(cell1.y - cell2.y);
-        // Adjacent if within 1 cell in both directions (8-directional adjacency)
-        return dx <= 1 && dy <= 1 && !(dx === 0 && dy === 0);
-      });
-    });
-  }
-
-  private getAdjacentPositionsForPiece(piece: GamePiece): Position[] {
-    if (!piece.position) return [];
-    
-    // Get all cells occupied by this piece
-    const occupiedCells = piece.shape.map(offset => ({
-      x: piece.position!.x + offset.x,
-      y: piece.position!.y + offset.y
-    }));
-    
-    // Get all adjacent positions (8-directional) for each occupied cell
-    const adjacentPositions = new Set<string>();
-    
-    occupiedCells.forEach(cell => {
-      // Add all 8 adjacent positions for this cell
-      for (let dx = -1; dx <= 1; dx++) {
-        for (let dy = -1; dy <= 1; dy++) {
-          if (dx === 0 && dy === 0) continue; // Skip the cell itself
-          
-          const adjPos = { x: cell.x + dx, y: cell.y + dy };
-          
-          // Only add if it's not already occupied by our own piece
-          if (!occupiedCells.some(occupied => occupied.x === adjPos.x && occupied.y === adjPos.y)) {
-            adjacentPositions.add(`${adjPos.x},${adjPos.y}`);
-          }
-        }
-      }
-    });
-    
-    // Convert back to Position objects
-    return Array.from(adjacentPositions).map(posStr => {
-      const [x, y] = posStr.split(',').map(Number);
-      return { x, y };
-    });
-  }
-
-
-  private calculatePieceStats(piece: GamePiece, allPieces: GamePiece[]): { attack: number; health: number; speed: number } {
-    if (!piece.position) {
-      return { attack: piece.stats.attack, health: piece.stats.health, speed: piece.stats.speed };
-    }
-
-    let attackBonus = 0;
-    let healthBonus = 0;
-    let speedBonus = 0;
-
-    // Find adjacent pieces using the proper adjacency check
-    const adjacentPieces = allPieces.filter(p => this.aretwoPiecesAdjacent(piece, p));
-
-    // Apply adjacency bonuses from plants and consumables to fish
-    adjacentPieces.forEach(adjacentPiece => {
-      if ((adjacentPiece.type === 'plant' || adjacentPiece.type === 'consumable') && piece.type === 'fish') {
-        const bonusAttack = adjacentPiece.attackBonus || 0;
-        const bonusHealth = adjacentPiece.healthBonus || 0;
-        const bonusSpeed = adjacentPiece.speedBonus || 0;
-        
-        attackBonus += bonusAttack;
-        healthBonus += bonusHealth;
-        speedBonus += bonusSpeed;
-      }
-    });
-
-    // Schooling fish bonuses
-    if (piece.tags.includes('schooling')) {
-      const adjacentSchoolingCount = adjacentPieces.filter(p => p.tags.includes('schooling')).length;
-      
-      if (piece.name === 'Neon Tetra') {
-        attackBonus += adjacentSchoolingCount;
-      }
-      
-      if (piece.name === 'Cardinal Tetra') {
-        attackBonus += adjacentSchoolingCount * 2;
-      }
-
-      // Double speed if 3+ schooling fish adjacent
-      if (adjacentSchoolingCount >= 3) {
-        speedBonus += piece.stats.speed;
-      }
-    }
-
-    // Include permanent bonuses from consumables
-    const permanentAttack = piece.permanentBonuses?.attack || 0;
-    const permanentHealth = piece.permanentBonuses?.health || 0;
-    const permanentSpeed = piece.permanentBonuses?.speed || 0;
-
-    return {
-      attack: piece.stats.attack + attackBonus + permanentAttack,
-      health: piece.stats.health + healthBonus + permanentHealth,
-      speed: piece.stats.speed + speedBonus + permanentSpeed
-    };
-  }
-
-
-
-
-  private findOptimalConsumablePosition(tank: Tank, consumable: GamePiece): Position | null {
-    let bestPosition: Position | null = null;
-    let maxFishTouched = 0;
-    
-    console.log(`🤖 Finding optimal position for consumable ${consumable.name}`);
-    
-    // Try every possible position and count adjacent fish
-    for (let y = 0; y < 6; y++) {
-      for (let x = 0; x < 8; x++) {
-        const position = { x, y };
-        
-        // Check if this position is valid for placement
-        if (!this.isValidPosition(tank, consumable, position)) {
-          continue;
-        }
-        
-        // Count how many fish this position would touch
-        const fishTouched = this.countAdjacentFish(tank, consumable, position);
-        
-        if (fishTouched > maxFishTouched) {
-          maxFishTouched = fishTouched;
-          bestPosition = position;
-        }
-      }
-    }
-    
-    console.log(`🤖 Best consumable position touches ${maxFishTouched} fish at ${bestPosition ? `(${bestPosition.x},${bestPosition.y})` : 'none'}`);
-    return bestPosition;
-  }
-
-  private countAdjacentFish(tank: Tank, consumable: GamePiece, position: Position): number {
-    const consumablePositions = consumable.shape.map(offset => ({
-      x: position.x + offset.x,
-      y: position.y + offset.y
-    }));
-    
-    // Get all positions adjacent to the consumable
-    const adjacentPositions = new Set<string>();
-    
-    consumablePositions.forEach(consPos => {
-      // Add all 8 adjacent positions for this consumable cell
-      for (let dx = -1; dx <= 1; dx++) {
-        for (let dy = -1; dy <= 1; dy++) {
-          if (dx === 0 && dy === 0) continue; // Skip the consumable cell itself
-          
-          const adjPos = { x: consPos.x + dx, y: consPos.y + dy };
-          
-          // Only add if it's within bounds and not occupied by the consumable itself
-          if (adjPos.x >= 0 && adjPos.x < 8 && adjPos.y >= 0 && adjPos.y < 6) {
-            const isConsumableCell = consumablePositions.some(cp => cp.x === adjPos.x && cp.y === adjPos.y);
-            if (!isConsumableCell) {
-              adjacentPositions.add(`${adjPos.x},${adjPos.y}`);
-            }
-          }
-        }
-      }
-    });
-    
-    // Count how many fish are at these adjacent positions
-    let fishCount = 0;
-    const adjacentPosArray = Array.from(adjacentPositions).map(posStr => {
-      const [x, y] = posStr.split(',').map(Number);
-      return { x, y };
-    });
-    
-    for (const adjPos of adjacentPosArray) {
-      // Find if there's a fish piece at this position
-      const fishAtPosition = tank.pieces.find(piece => {
-        if (piece.type !== 'fish' || !piece.position) return false;
-        
-        // Check if any part of this fish occupies the adjacent position
-        return piece.shape.some(offset => {
-          const fishCellX = piece.position!.x + offset.x;
-          const fishCellY = piece.position!.y + offset.y;
-          return fishCellX === adjPos.x && fishCellY === adjPos.y;
-        });
-      });
-      
-      if (fishAtPosition) {
-        fishCount++;
-      }
-    }
-    
-    return fishCount;
-  }
-
-
-
-  private findValidPositionForOpponent(tank: Tank, piece: GamePiece): Position | null {
-    // Try to find a valid position for the piece without overlapping existing pieces
-    for (let y = 0; y < 6; y++) {
-      for (let x = 0; x < 8; x++) {
-        const position = { x, y };
-        if (this.isValidPositionForNewPiece(tank, piece, position)) {
-          return position;
-        }
-      }
-    }
-    return null; // No valid position found
-  }
-
-  private isValidPositionForNewPiece(tank: Tank, piece: GamePiece, position: Position): boolean {
-    // For new pieces, simply check if grid cells are empty
-    for (const offset of piece.shape) {
-      const x = position.x + offset.x;
-      const y = position.y + offset.y;
-      
-      if (x < 0 || x >= 8 || y < 0 || y >= 6) {
-        return false;
-      }
-      
-      // For new pieces, any occupied cell means invalid position
-      if (tank.grid[y][x] !== null) {
-        return false;
-      }
-    }
-    
-    return true;
-  }
-
-  private isValidPositionForGrid(grid: (string | null)[][], piece: GamePiece, position: Position): boolean {
-    for (const offset of piece.shape) {
-      const x = position.x + offset.x;
-      const y = position.y + offset.y;
-      
-      if (x < 0 || x >= 8 || y < 0 || y >= 6) {
-        return false;
-      }
-      
-      if (grid[y][x] !== null) {
-        return false;
-      }
-    }
-    
-    return true;
-  }
-
-
 
   private async runBattleSimulation(socketId: string): Promise<void> {
     const gameState = await this.getSession(socketId);
