+++ game-engine/src/tank/tank.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { TankService } from './tank.service';
+
+@Module({
+  providers: [TankService],
+  exports: [TankService],
+})
+export class TankModule {}
